+++ src/DotRecast.Detour/DtPolyDetail.cs
@@ -25,10 +25,10 @@ namespace DotRecast.Detour
     {
         public readonly int vertBase; //< The offset of the vertices in the dtMeshTile::detailVerts array.
         public readonly int triBase; //< The offset of the triangles in the dtMeshTile::detailTris array.
-        public readonly int vertCount; //< The number of vertices in the sub-mesh.
-        public readonly int triCount; //< The number of triangles in the sub-mesh.
+        public readonly byte vertCount; //< The number of vertices in the sub-mesh.
+        public readonly byte triCount; //< The number of triangles in the sub-mesh.
 
-        public DtPolyDetail(int vertBase, int triBase, int vertCount, int triCount)
+        public DtPolyDetail(int vertBase, int triBase, byte vertCount, byte triCount)
         {
             this.vertBase = vertBase;
             this.triBase = triBase;