+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Changed to consolidate vector-related functions into one place.
 - Changed stack handling from List to a fixed-size array with manual index management for optimization in `RcLayers.BuildHeightfieldLayers()`
 - Changed to use Span<byte> and stackalloc for improved performance and memory management in `RcLayers.BuildHeightfieldLayers()`
+- Changed vertCount and triCount to byte in `DtPolyDetail`
 
 ### Removed
 - Removed RcMeshDetails.VdistSq2(float[], float[])